+++ packages/lines/package.json
@@ -1,6 +1,10 @@
 {
   "name": "@oasis-engine-toolkit/lines",
   "version": "0.8.0-alpha.10",
+  "publishConfig": {
+    "access": "public",
+    "registry": "https://registry.npmjs.org"
+  },
   "license": "MIT",
   "scripts": {
     "b:types": "tsc",
