+++ packages/lines/src/line/material/DashMaterial.ts
@@ -1,118 +1,10 @@
-import { Shader, Engine, Texture2D, Vector2 } from "oasis-engine";
+import { Shader, Engine } from "oasis-engine";
 import { LineMaterial } from "./LineMaterial";
-
-//-- Shader 代码
-const vertexSource = `
-attribute vec2 a_pos;
-attribute vec2 a_normal;
-attribute vec2 a_data;
-attribute float a_lengthsofar;
-
-uniform mat4 u_MVPMat;
-uniform float u_width;
-uniform vec2 u_dash;
-
-varying vec2 v_origin;
-varying vec2 v_position;
-varying float v_direction;
-varying float v_part;
-varying vec2 v_tex;
-
-void main() {
-    v_direction = a_data.x;
-    v_part = a_data.y;
-    float layer_index = 1.0;
-
-
-    v_origin = a_pos;
-
-    float texcoord_y = 0.0;
-
-    texcoord_y = a_lengthsofar / (u_dash.x + u_dash.y);
-    if (v_direction == 1.0) {
-        v_tex = vec2(1.0, texcoord_y);
-    } else {
-        v_tex = vec2(0.0, texcoord_y);
-    }
-    vec2 position = a_pos + a_normal * u_width;
-    v_position = position;
-    gl_Position = u_MVPMat * vec4(position, 0.0, 1);
-}
-  `;
-
-const fragmentSource = `
-precision highp float;
-
-uniform vec4 u_color;
-uniform int u_join;
-uniform int u_cap;
-uniform float u_width;
-uniform sampler2D u_texture;
-
-varying vec2 v_origin;
-varying vec2 v_position;
-varying float v_direction;
-varying float v_part;
-varying vec2 v_tex;
-
-float IS_CAP = 0.0;
-
-void main() {
-    vec4 finalColor;
-    if (u_cap == 0 && v_part == IS_CAP) {
-      if (distance(v_position, v_origin) > u_width) {
-        discard;
-      }
-    }
-    if (u_join == 1 && v_part > 1.5) {
-      if (distance(v_position, v_origin) > u_width) {
-        discard;
-      }
-    }
-    vec4 textureColor = texture2D(u_texture, v_tex);
-    if (textureColor.a <= 0.5) {
-      gl_FragColor = vec4(u_color.rgb, 0.0);
-    } else {
-      gl_FragColor = u_color;
-    }
-}
-`;
-
-Shader.create("dash", vertexSource, fragmentSource);
+import "./dashShader";
 
 export class DashMaterial extends LineMaterial {
-  private _dash: Vector2;
-
-  /**
-   * The dash sequence is a series of on/off lengths in points.
-   */
-  get dash() {
-    return this._dash;
-  }
-
-  set dash(val: Vector2) {
-    this._dash = val;
-    this.shaderData.setVector2("u_dash", val);
-    const texture = new Texture2D(this.engine, 1, Math.ceil((val.x + val.y) * 10));
-    texture.setPixelBuffer(this._generateDashTexture(val));
-    this.shaderData.setTexture("u_texture", texture);
-  }
-
   constructor(engine: Engine) {
     super(engine);
     this.shader = Shader.find("dash");
-  }
-
-  private _generateDashTexture(dash) {
-    const pixels: number[] = [];
-    const length = Math.ceil((dash.x + dash.y) * 10);
-    for (let index = 0; index < length; index++) {
-      if (index < dash.x * 10) {
-        pixels.push(255, 255, 255, 255);
-      } else {
-        pixels.push(255, 255, 255, 0);
-      }
-    }
-    return new Uint8Array(pixels);
   }
 }
